+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## Unreleased
+
+### Fixed
+- Fixed an error that could occur when running `command` tasks with a PHP binary or project path containing spaces, such as under Laravel Herd ([#1](https://github.com/jalendport/craft-metronome/issues/1))
+
 ## 1.0.0 - 2026-07-21
 
 ### Added